+++ projects/abuild/test/settings_test.cpp
@@ -1,5 +1,5 @@
 import abuild;
-import abuild_test_cache;
+import abuild_test_utilities;
 
 using atest::assert_;
 using atest::assert_fail;
